# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Verifyflow API
  version: 0.5.0
  description: |
    Public HTTP contract for the registered health, authentication, account, rental, payment,
    messaging, webhook, and admin routes. All customer request bodies are JSON. Responses include an `X-Request-ID` header.
    Access tokens are short-lived JWTs; refresh tokens are opaque, single-use tokens that rotate
    on refresh.

    Rental routes require a verified email address in addition to a valid access token. Admin
    routes require the signed-in user's role to be `admin`; every account starts as `user`.

    Twilio, Vonage, and Telnyx are supported behind a provider-neutral interface. Inbound provider
    callbacks require their vendor signature and are persisted idempotently.
servers:
  - url: /
    description: Same origin as this documentation page
...
paths:
...
  /api/v1/numbers:
    get:
      tags: [Rentals]
      summary: Search available numbers
      description: Lists provider numbers currently AVAILABLE for rental, optionally filtered by type and capability.
      operationId: searchNumbers
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [Local, TollFree, Mobile]
        - name: sms
          in: query
          description: When true, only return numbers with SMS capability.
          schema:
            type: boolean
        - name: mms
          in: query
          description: When true, only return numbers with MMS capability.
          schema:
            type: boolean
        - name: voice
          in: query
          description: When true, only return numbers with voice capability.
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: Opaque pagination cursor from a previous page's next_cursor.
          schema:
            type: string
      responses:
        '200':
          description: Page of available numbers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NumbersPage'
        '400':
          description: Invalid type or limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_request
                  message: Request is missing or has invalid fields.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmailNotVerified'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
...
  /api/v1/orders/{id}:
    get:
      tags: [Rentals]
      summary: Get an order owned by the signed-in user
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          description: The order.
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmailNotVerified'
        '404':
          description: No such order, or it belongs to another user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: not_found
                  message: The requested order was not found.
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    NumberSummary:
      type: object
      required: [id, phone_number, number_type, sms_enabled, mms_enabled, voice_enabled]
      properties:
        id:
          type: string
          format: uuid
        phone_number:
          type: string
          example: '+14155550001'
        number_type:
          type: string
          enum: [Local, TollFree, Mobile]
        sms_enabled:
          type: boolean
        mms_enabled:
          type: boolean
        voice_enabled:
          type: boolean
    NumbersPage:
      type: object
      required: [numbers]
      properties:
        numbers:
          type: array
          items:
            $ref: '#/components/schemas/NumberSummary'
        next_cursor:
          type: string
          description: Opaque cursor for the next page; empty when there are no more results.
...
    Order:
      type: object
      required:
        - id
        - rental_id
        - provider_number_id
        - rental_plan_id
        - status
        - plan_code
        - plan_name
        - duration_seconds
        - price_minor_units
        - currency
        - reservation_expires_at
        - created_at
      properties:
        id:
          type: string
          format: uuid
        rental_id:
          type: string
          format: uuid
        provider_number_id:
          type: string
          format: uuid
        rental_plan_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, EXPIRED]
        plan_code:
          type: string
        plan_name:
          type: string
        duration_seconds:
          type: integer
        price_minor_units:
          type: integer
          format: int64
          description: Server-priced snapshot at reservation time; unaffected by later plan-price changes.
        currency:
          type: string
        reservation_expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid access token.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer realm="migo"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: Authentication is required.
    EmailNotVerified:
      description: The signed-in user's email address is not verified.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: email_not_verified
              message: Please verify your email address to use this feature.
...
    InternalError:
      description: Unexpected server error. Details are not exposed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: internal_error
              message: Something went wrong.
    Timeout:
      description: Request exceeded its deadline.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: timeout
              message: The request timed out.